implement compare smartphone & refactor classes
