+++ model/Recommendation.go
@@ -1,0 +1,14 @@
+package model
+
+type RecommendationsRequest struct {
+	Price     string `json:"price"`
+	Processor string `json:"processor"`
+	Camera    string `json:"camera"`
+	Battery   string `json:"battery"`
+	Ram       string `json:"ram"`
+	Storage   string `json:"storage"`
+}
+
+type RecommendationsResponse struct {
+	Recommendations []string `json:"recommendations"`
+}
