+++ model/Smartphone.go
@@ -1,22 +1,9 @@
 package model
 
-type RecommendationsRequest struct {
-	Price     string `json:"price"`
-	Processor string `json:"processor"`
-	Camera    string `json:"camera"`
-	Battery   string `json:"battery"`
-	Ram       string `json:"ram"`
-	Storage   string `json:"storage"`
-}
-
 // Struct to hold an item and its similarity score
 type SmartphoneSimilarity struct {
 	Name       string
 	Similarity float64
-}
-
-type RecommendationsResponse struct {
-	Recommendations []string `json:"recommendations"`
 }
 
 type Smartphone struct {
